Refactor move ResourceInterface and WorldInterface
